+++ pkg/helpers/shared.go
@@ -1,0 +1,99 @@
+package helpers
+
+import (
+	"context"
+	"crypto/ecdsa"
+	"crypto/elliptic"
+	"crypto/rand"
+	"crypto/x509"
+	"crypto/x509/pkix"
+	"encoding/pem"
+	"fmt"
+	"math/big"
+	math_rand "math/rand"
+	"os"
+	"time"
+
+	"github.com/qernal/cli-qernal/commands/auth"
+	"github.com/qernal/cli-qernal/pkg/client"
+)
+
+func GenerateSelfSignedCert() ([]byte, []byte, error) {
+	// Generate a new ECDSA private key
+	privateKey, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	// Create a self-signed certificate template
+	template := x509.Certificate{
+		SerialNumber: big.NewInt(1),
+		Subject: pkix.Name{
+			Organization: []string{"Example Corp"},
+		},
+		NotBefore: time.Now(),
+		NotAfter:  time.Now().Add(365 * 24 * time.Hour), // Valid for 1 year
+		KeyUsage:  x509.KeyUsageKeyEncipherment | x509.KeyUsageDigitalSignature,
+		ExtKeyUsage: []x509.ExtKeyUsage{
+			x509.ExtKeyUsageServerAuth,
+		},
+		BasicConstraintsValid: true,
+	}
+
+	// Create the self-signed certificate
+	derBytes, err := x509.CreateCertificate(rand.Reader, &template, &template, &privateKey.PublicKey, privateKey)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	// Encode the public key (certificate) to PEM
+	certPEM := pem.EncodeToMemory(&pem.Block{
+		Type:  "CERTIFICATE",
+		Bytes: derBytes,
+	})
+
+	// Encode the private key to PEM
+	privateKeyBytes, err := x509.MarshalECPrivateKey(privateKey)
+	if err != nil {
+		return nil, nil, err
+	}
+	privateKeyPEM := pem.EncodeToMemory(&pem.Block{
+		Type:  "EC PRIVATE KEY",
+		Bytes: privateKeyBytes,
+	})
+
+	return certPEM, privateKeyPEM, nil
+}
+
+func FetchDek(projectID string) (string, int32, error) {
+	ctx := context.Background()
+	token, err := auth.GetQernalToken()
+	if err != nil {
+		return "", 0, err
+	}
+
+	qc, err := client.New(ctx, nil, nil, token)
+	if err != nil {
+		return "", 0, err
+	}
+
+	resp, r, err := qc.SecretsAPI.ProjectsSecretsGet(context.Background(), projectID, "dek").Execute()
+
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Error when calling `ProjectsAPI.ProjectsSecretsGet``: %v\n", err)
+		fmt.Fprintf(os.Stderr, "Full HTTP response: %v\n", r)
+
+		return "", 0, err
+	}
+
+	return resp.Payload.SecretMetaResponseDek.Public, resp.Revision, nil
+}
+
+func RandomSecretName() string {
+	const charset = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
+	b := make([]byte, 8)
+	for i := range b {
+		b[i] = charset[math_rand.Intn(len(charset))]
+	}
+	return fmt.Sprintf("TERRA_%s", string(b))
+}
